refactor: Dao 객체의 불필요한 인터페이스 제거
